modularizacion hecha
