+++ Django/homebanking/clientes/serializers.py
@@ -1,11 +1,8 @@
 
-from dataclasses import field
+
 from .models import User, Direccion
 from rest_framework import serializers
-from cuentas.models import Cuenta
-from prestamos.models import Prestamo, Sucursal
-from tarjetas.models import Tarjeta
-from homebank.models import SolicitudesPrestamos
+
 
 #item 1
 class UserClienteSerializer(serializers.ModelSerializer):
@@ -15,29 +12,29 @@ class UserClienteSerializer(serializers.ModelSerializer):
         read_only_fields = ("id", "date_joined", "is_staff")
         
 
-#item 2
-class CuentaSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Cuenta
-        fields = ["cuentaid","balance"]
+# #item 2
+# class CuentaSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Cuenta
+#         fields = ["cuentaid","balance"]
        
-#item 3 
-class PrestamoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Prestamo
-        fields = ["loan_type", "loan_total"]
+# #item 3 
+# class PrestamoSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Prestamo
+#         fields = ["loan_type", "loan_total"]
        
-#item 5 
-class TarjetasSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tarjeta
-        fields = "__all__"
+# #item 5 
+# class TarjetasSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Tarjeta
+#         fields = "__all__"
 
-#item 6 y 7
-class SolicitudesPrestamosSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = SolicitudesPrestamos
-        fields = "__all__"
+# #item 6 y 7
+# class SolicitudesPrestamosSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = SolicitudesPrestamos
+#         fields = "__all__"
 
 #item 8
 class DireccionSerializer(serializers.ModelSerializer):
@@ -47,8 +44,8 @@ class DireccionSerializer(serializers.ModelSerializer):
         
 #item 9
 
-class SucursalSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Sucursal
-        fields = "__all__"
-        read_only_fields = ("branch_id", "branch_number", "branch_name", "branch_address_id")
+# class SucursalSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Sucursal
+#         fields = "__all__"
+#         read_only_fields = ("branch_id", "branch_number", "branch_name", "branch_address_id")
