+++ Django/homebanking/cuentas/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+from .views import  CuentaDetail
+
+urlpatterns = [
+    path("cuenta/", CuentaDetail.as_view()),
+    ]
