+++ Django/homebanking/cuentas/views.py
@@ -1,10 +1,13 @@
+
 from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status, permissions
+from rest_framework import status, permissions, generics, viewsets
 from .serializers import CuentaSerializer
-from .models import Cuenta
+from cuentas.models import Cuenta
 
+
+#item 2
 class CuentaDetail(APIView):
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
     def get(self, request):
@@ -15,3 +18,5 @@ class CuentaDetail(APIView):
         if usuario:
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(status=status.HTTP_400_BAD_REQUEST)
+    
+
