+++ Django/homebanking/homebank/views.py
@@ -1,11 +1,13 @@
-from django.shortcuts import render, redirect, HttpResponse
+from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.urls import reverse
 from clientes.models import Cliente
-
 from homebank.models import SolicitudesPrestamos
-
 from .forms import SolicitudPrestamo
+from django.shortcuts import render
+from rest_framework import permissions, viewsets
+from .serializers import SolicitudesPrestamosSerializer
+from homebank.models import SolicitudesPrestamos
 
 @login_required
 def vista(request):
@@ -40,3 +42,10 @@ def formularioSolicitud(request):
                 
                     
     return render(request,"homebank/prestamos.html", {"form":formulario_solicitud})
+
+
+#item 6 y 7
+class SolicitudesPrestamosViewset(viewsets.ModelViewSet):
+    queryset = SolicitudesPrestamos.objects.all()
+    serializer_class = SolicitudesPrestamosSerializer
+    permission_classes = [permissions.IsAdminUser]
