+++ Django/homebanking/tarjetas/serializers.py
@@ -1,0 +1,10 @@
+
+from rest_framework import serializers
+from tarjetas.models import Tarjeta
+
+
+#item 5 
+class TarjetasSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tarjeta
+        fields = "__all__"
